pyproject.toml
[project]
name = "sdr"
version = "0.1.0"
description = "Sales Development Representative Agent System - Multi-agent system for generating and sending cold sales emails"
authors = [{ name = "SDR Team" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "openai-agents>=0.0.15",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
    "certifi>=2024.0.0",
]

[project.scripts]
sdr = "sdr.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sdr"]

